fix: remove bassmix and fix a lot warnings

+++ Engine/src/Misc/md5.cpp
@@ -5,47 +5,47 @@
 
 #include <Misc/md5.h>
 
- /*
-  * Constants defined by the MD5 algorithm
-  */
+/*
+ * Constants defined by the MD5 algorithm
+ */
 #define A 0x67452301
 #define B 0xefcdab89
 #define C 0x98badcfe
 #define D 0x10325476
 
 static uint32_t S[] = { 7, 12, 17, 22, 7, 12, 17, 22, 7, 12, 17, 22, 7, 12, 17, 22,
-                       5,  9, 14, 20, 5,  9, 14, 20, 5,  9, 14, 20, 5,  9, 14, 20,
-                       4, 11, 16, 23, 4, 11, 16, 23, 4, 11, 16, 23, 4, 11, 16, 23,
-                       6, 10, 15, 21, 6, 10, 15, 21, 6, 10, 15, 21, 6, 10, 15, 21 };
+                        5, 9, 14, 20, 5, 9, 14, 20, 5, 9, 14, 20, 5, 9, 14, 20,
+                        4, 11, 16, 23, 4, 11, 16, 23, 4, 11, 16, 23, 4, 11, 16, 23,
+                        6, 10, 15, 21, 6, 10, 15, 21, 6, 10, 15, 21, 6, 10, 15, 21 };
 
 static uint32_t K[] = { 0xd76aa478, 0xe8c7b756, 0x242070db, 0xc1bdceee,
-                       0xf57c0faf, 0x4787c62a, 0xa8304613, 0xfd469501,
-                       0x698098d8, 0x8b44f7af, 0xffff5bb1, 0x895cd7be,
-                       0x6b901122, 0xfd987193, 0xa679438e, 0x49b40821,
-                       0xf61e2562, 0xc040b340, 0x265e5a51, 0xe9b6c7aa,
-                       0xd62f105d, 0x02441453, 0xd8a1e681, 0xe7d3fbc8,
-                       0x21e1cde6, 0xc33707d6, 0xf4d50d87, 0x455a14ed,
-                       0xa9e3e905, 0xfcefa3f8, 0x676f02d9, 0x8d2a4c8a,
-                       0xfffa3942, 0x8771f681, 0x6d9d6122, 0xfde5380c,
-                       0xa4beea44, 0x4bdecfa9, 0xf6bb4b60, 0xbebfbc70,
-                       0x289b7ec6, 0xeaa127fa, 0xd4ef3085, 0x04881d05,
-                       0xd9d4d039, 0xe6db99e5, 0x1fa27cf8, 0xc4ac5665,
-                       0xf4292244, 0x432aff97, 0xab9423a7, 0xfc93a039,
-                       0x655b59c3, 0x8f0ccc92, 0xffeff47d, 0x85845dd1,
-                       0x6fa87e4f, 0xfe2ce6e0, 0xa3014314, 0x4e0811a1,
-                       0xf7537e82, 0xbd3af235, 0x2ad7d2bb, 0xeb86d391 };
+                        0xf57c0faf, 0x4787c62a, 0xa8304613, 0xfd469501,
+                        0x698098d8, 0x8b44f7af, 0xffff5bb1, 0x895cd7be,
+                        0x6b901122, 0xfd987193, 0xa679438e, 0x49b40821,
+                        0xf61e2562, 0xc040b340, 0x265e5a51, 0xe9b6c7aa,
+                        0xd62f105d, 0x02441453, 0xd8a1e681, 0xe7d3fbc8,
+                        0x21e1cde6, 0xc33707d6, 0xf4d50d87, 0x455a14ed,
+                        0xa9e3e905, 0xfcefa3f8, 0x676f02d9, 0x8d2a4c8a,
+                        0xfffa3942, 0x8771f681, 0x6d9d6122, 0xfde5380c,
+                        0xa4beea44, 0x4bdecfa9, 0xf6bb4b60, 0xbebfbc70,
+                        0x289b7ec6, 0xeaa127fa, 0xd4ef3085, 0x04881d05,
+                        0xd9d4d039, 0xe6db99e5, 0x1fa27cf8, 0xc4ac5665,
+                        0xf4292244, 0x432aff97, 0xab9423a7, 0xfc93a039,
+                        0x655b59c3, 0x8f0ccc92, 0xffeff47d, 0x85845dd1,
+                        0x6fa87e4f, 0xfe2ce6e0, 0xa3014314, 0x4e0811a1,
+                        0xf7537e82, 0xbd3af235, 0x2ad7d2bb, 0xeb86d391 };
 
 /*
  * Padding used to make the size (in bits) of the input congruent to 448 mod 512
  */
 static uint8_t PADDING[] = { 0x80, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-                            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 };
+                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                             0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 };
 
 /*
  * Bit-manipulation functions defined by the MD5 algorithm
@@ -55,18 +55,19 @@ static uint8_t PADDING[] = { 0x80, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
 #define H(X, Y, Z) (X ^ Y ^ Z)
 #define I(X, Y, Z) (Y ^ (X | ~Z))
 
- /*
-  * Rotates a 32-bit word left by n bits
-  */
-uint32_t rotateLeft(uint32_t x, uint32_t n) {
+/*
+ * Rotates a 32-bit word left by n bits
+ */
+uint32_t rotateLeft(uint32_t x, uint32_t n)
+{
     return (x << n) | (x >> (32 - n));
 }
-
 
 /*
  * Initialize a context
  */
-void md5Init(MD5Context* ctx) {
+void md5Init(MD5Context *ctx)
+{
     ctx->size = (uint64_t)0;
 
     ctx->buffer[0] = (uint32_t)A;
@@ -81,8 +82,9 @@ void md5Init(MD5Context* ctx) {
  * If the input fills out a block of 512 bits, apply the algorithm (md5Step)
  * and save the result in the buffer. Also updates the overall size.
  */
-void md5Update(MD5Context* ctx, uint8_t* input_buffer, size_t input_len) {
-    uint32_t input[16];
+void md5Update(MD5Context *ctx, uint8_t *input_buffer, size_t input_len)
+{
+    uint32_t     input[16];
     unsigned int offset = ctx->size % 64;
     ctx->size += (uint64_t)input_len;
 
@@ -100,9 +102,9 @@ void md5Update(MD5Context* ctx, uint8_t* input_buffer, size_t input_len) {
                 // The local variable `input` our 512-bit chunk separated into 32-bit words
                 // we can use in calculations
                 input[j] = (uint32_t)(ctx->input[(j * 4) + 3]) << 24 |
-                    (uint32_t)(ctx->input[(j * 4) + 2]) << 16 |
-                    (uint32_t)(ctx->input[(j * 4) + 1]) << 8 |
-                    (uint32_t)(ctx->input[(j * 4)]);
+                           (uint32_t)(ctx->input[(j * 4) + 2]) << 16 |
+                           (uint32_t)(ctx->input[(j * 4) + 1]) << 8 |
+                           (uint32_t)(ctx->input[(j * 4)]);
             }
             md5Step(ctx->buffer, input);
             offset = 0;
@@ -114,8 +116,9 @@ void md5Update(MD5Context* ctx, uint8_t* input_buffer, size_t input_len) {
  * Pad the current input to get to 448 bytes, append the size in bits to the very end,
  * and save the result of the final iteration into digest.
  */
-void md5Finalize(MD5Context* ctx) {
-    uint32_t input[16];
+void md5Finalize(MD5Context *ctx)
+{
+    uint32_t     input[16];
     unsigned int offset = ctx->size % 64;
     unsigned int padding_length = offset < 56 ? 56 - offset : (56 + 64) - offset;
 
@@ -127,9 +130,9 @@ void md5Finalize(MD5Context* ctx) {
     // Last two 32-bit words are the two halves of the size (converted from bytes to bits)
     for (unsigned int j = 0; j < 14; ++j) {
         input[j] = (uint32_t)(ctx->input[(j * 4) + 3]) << 24 |
-            (uint32_t)(ctx->input[(j * 4) + 2]) << 16 |
-            (uint32_t)(ctx->input[(j * 4) + 1]) << 8 |
-            (uint32_t)(ctx->input[(j * 4)]);
+                   (uint32_t)(ctx->input[(j * 4) + 2]) << 16 |
+                   (uint32_t)(ctx->input[(j * 4) + 1]) << 8 |
+                   (uint32_t)(ctx->input[(j * 4)]);
     }
     input[14] = (uint32_t)(ctx->size * 8);
     input[15] = (uint32_t)((ctx->size * 8) >> 32);
@@ -148,7 +151,8 @@ void md5Finalize(MD5Context* ctx) {
 /*
  * Step on 512 bits of input with the main MD5 algorithm.
  */
-void md5Step(uint32_t* buffer, uint32_t* input) {
+void md5Step(uint32_t *buffer, uint32_t *input)
+{
     uint32_t AA = buffer[0];
     uint32_t BB = buffer[1];
     uint32_t CC = buffer[2];
@@ -195,29 +199,31 @@ void md5Step(uint32_t* buffer, uint32_t* input) {
  * Functions that run the algorithm on the provided input and put the digest into result.
  * result should be able to store 16 bytes.
  */
-void md5String(char* input, uint8_t* result) {
+void md5String(char *input, uint8_t *result)
+{
     MD5Context ctx;
     md5Init(&ctx);
-    md5Update(&ctx, (uint8_t*)input, strlen(input));
+    md5Update(&ctx, (uint8_t *)input, strlen(input));
     md5Finalize(&ctx);
 
     memcpy(result, ctx.digest, 16);
 }
 
-void md5Buffer(char* input, size_t input_len, uint8_t* result) {
+void md5Buffer(char *input, size_t input_len, uint8_t *result)
+{
     MD5Context ctx;
     md5Init(&ctx);
 
     const int maxSize = 1024;
-    int offset = 0;
+    int       offset = 0;
 
     while (offset < input_len) {
-        int size = input_len - offset;
+        int size = (int)input_len - offset;
         if (size > maxSize) {
             size = maxSize;
         }
 
-        md5Update(&ctx, (uint8_t*)(input + offset), size);
+        md5Update(&ctx, (uint8_t *)(input + offset), size);
         offset += size;
     }
 
@@ -226,15 +232,16 @@ void md5Buffer(char* input, size_t input_len, uint8_t* result) {
     memcpy(result, ctx.digest, 16);
 }
 
-void md5File(FILE* file, uint8_t* result) {
-    char* input_buffer = (char*)malloc(1024);
+void md5File(FILE *file, uint8_t *result)
+{
+    char  *input_buffer = (char *)malloc(1024);
     size_t input_size = 0;
 
     MD5Context ctx;
     md5Init(&ctx);
 
     while ((input_size = fread(input_buffer, 1, 1024, file)) > 0) {
-        md5Update(&ctx, (uint8_t*)input_buffer, input_size);
+        md5Update(&ctx, (uint8_t *)input_buffer, input_size);
     }
 
     md5Finalize(&ctx);
